Split interaction audit Playwright helpers
